Была попытка сделать u64 на чистом C89. Решил использовать long в случае отстутствия компилятором поддержки long long, иначе, даже если и сделаю софтовое умножение и делени, то это будет оч. медленно. Коммит, если вдруг захочу опять этим занятся.

+++ examples/main.cpp
@@ -8,7 +8,7 @@ std::ifstream file("examples/asm/perf_test.ncvm", std::ios::binary);
 
 
 
-const unsigned char* get_next_n_bytes(unsigned long long n, void* data_p) {
+const unsigned char* get_next_n_bytes(unsigned long n, void* data_p) {
     unsigned char* data = new unsigned char[n];
     file.read((char*)data, n);
     return data;
